+++ actix-protobuf/CHANGES.md
@@ -4,6 +4,10 @@
 
 * Minimum supported Rust version(MSRV) is now 1.40.0.
 
+## unreleased
+
+* Minimize `futures` dependency
+
 ## 0.5.1 (2019-02-17)
 
 * Move repository to actix-extras